Extract assignations
